fix(geometry): gate facet_weld's local-frame anchor to coordinates that need it
The watertightness census (run 31776201061) caught three real regressions
from the anchored plane-offset math: dental_clinic.ifc #1311 and
ISSUE_129...IGC_V17.ifc #81562 lost triangles (over-merged distinct facets),
and rvt01.ifc #6588 became newly triangulator-dependent.

The maintainer's hypothesis was that MAX_OFFSET_JITTER had become too
permissive at these hosts' coordinate magnitudes. Measured directly: none of
the three hosts are large-site-coordinate meshes (max|raw| < 30 m on every
one, well under LARGE_COORD_THRESHOLD_METERS), so a jitter-bound change
would not address the mechanism.

Root cause: `Facet::offset` is `n·(v - anchor)` using EACH FACET'S OWN
normal, not one shared reference normal. Comparing two facets' offsets
(the Step 4 clustering gap) therefore carries a residual
`(n_i - n_j)·anchor` cross-term the pre-anchor `n·v` formulation never had.
`(n_i - n_j)` is bounded only by the NORMAL_QUANT bucket width (~0.057deg)
and by normal-computation noise from anywhere upstream (measured
contributor on the regressed hosts: ~5e-5, from the void-cut local-frame
rotation in router/voids/mod.rs, two orders above plain f32 import jitter)
- and `anchor` is a whole-mesh bbox-min, so it can be metres from the
specific facet pair being compared even on a small local mesh. That
cross-term can swallow a genuine sub-millimetre plane separation
(over-merge) or inflate a genuinely-mergeable gap past MAX_OFFSET_JITTER in
a triangulation-dependent way - exactly the two failure shapes observed.

Fix: only use the bbox-min anchor once the mesh's own raw coordinates are
large enough that the amplification it exists to fix could actually bite
(ANCHOR_ENGAGE_METERS = 100.0, verified against both this file's own
large-site tests at ~150 m and ~5000-7000 m). Below that, anchor at the
origin, which reproduces the pre-anchor `n·v` formulation exactly
(bit-for-bit) - what the golden watertightness census was blessed against.
Verified against the census's own uploaded rows for all three regressed
hosts, under both triangulators: exact golden match (open/tris counts and
triangulator-invariance all restored), with the full geometry test suite
(644 lib tests + 86 integration binaries) green.

Also fixes a review finding (PR #2611, coderabbitai): `acc_n · x = acc_off`
is the accumulated (non-unit) plane equation; converting it to unit-normal
form requires dividing BOTH sides by the same scalar (`|acc_n|`), not
`acc_n` by `|acc_n|` while `acc_off` by `wsum` - those differ whenever
member facets' normals aren't identical, desyncing `pd` from `pn`.

Addresses the second review finding by strengthening
welds_offset_jitter_at_large_site_coordinates: the prior fixture used two
exactly axis-aligned triangles (identical Z-normal), so it never exercised
per-facet normal error at all. Replaced with a non-axis-aligned pair whose
normals differ after real f32 re-quantization, asserting the raw (pre-fix)
offset gap actually exceeds MAX_OFFSET_JITTER before asserting the anchored
weld recovers it.

Moves facet_weld's inline unit tests into the ratchet-exempt
facet_weld_scoped_tests.rs sibling (module_size_ratchet.rs's `_tests.rs`
exemption): facet_weld.rs was at the allowlist's 1229-line budget with zero
headroom; it is now 907 lines.
